declare the vlen member of tagged_value_t structures as size_t instead of uint32_t
as suggested in https://github.com/xant/libhl/issues/11

+++ src/linklist.c
@@ -833,7 +833,7 @@ list_create_tagged_value(char *tag, void *val, uint32_t vlen)
         else
         {
             newval->value = (void *)strdup((char *)val);
-            newval->vlen = (uint32_t)strlen((char *)val);
+            newval->vlen = strlen((char *)val);
             newval->type = TV_TYPE_STRING;
         }
     }
